// src/api/model/employeeResponse.ts
/**
 * OpenApi Spezifikation for GeekHR-Connect
 * API Dokumentation für die Spring Boot Anwendung
 *
 * Contact: asmitterand@yahoo.fr
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface EmployeeResponse { 
    id: string;
    gender: EmployeeResponse.GenderEnum;
    firstName: string;
    lastName: string;
    email: string;
    phone: string;
    dateOfBirth: string;
    city: string;
    country: string;
    remainingVacationDays: number;
    onVacation: boolean;
    position: EmployeeResponse.PositionEnum;
    imageURL: string;
    superiorName?: string;
    superiorId?: string;
}
export namespace EmployeeResponse {
    export type GenderEnum = 'MEN' | 'WOMEN';
    export const GenderEnum = {
        Men: 'MEN' as GenderEnum,
        Women: 'WOMEN' as GenderEnum
    };
    export type PositionEnum = 'CEO' | 'CTO' | 'COO' | 'TEAM_MANAGER_SOFTWARE' | 'SENIOR_SOFTWARE_DEVELOPER' | 'SOFTWARE_DEVELOPER' | 'JUNIOR_SOFTWARE_DEVELOPER' | 'WORKING_STUDENT';
    export const PositionEnum = {
        Ceo: 'CEO' as PositionEnum,
        Cto: 'CTO' as PositionEnum,
        Coo: 'COO' as PositionEnum,
        TeamManagerSoftware: 'TEAM_MANAGER_SOFTWARE' as PositionEnum,
        SeniorSoftwareDeveloper: 'SENIOR_SOFTWARE_DEVELOPER' as PositionEnum,
        SoftwareDeveloper: 'SOFTWARE_DEVELOPER' as PositionEnum,
        JuniorSoftwareDeveloper: 'JUNIOR_SOFTWARE_DEVELOPER' as PositionEnum,
        WorkingStudent: 'WORKING_STUDENT' as PositionEnum
    };
}